collect-copyright: fix parsing of gold link maps
In some cases (I don't know which), the gold linker might emit a link map
which looks like this:

    Archive member included to satisfy reference by file (symbol)

    [snip]
    libquick-lint-js-lib.a        libquick-lint-js-lib.a (_ZN13quick_lint_js16web_demo_locatorC1ENS_18padded_string_viewE)
    ../libboost.a                 CMakeFiles/quick-lint-js.dir/main.cpp.o (_ZN5boost9container3pmr19new_delete_resourceEv)
    ../libboost.a                 ../libboost.a (_ZN5boost9container25dlmalloc_global_sync_lockEv)
    ../libboost.a                 ../libboost.a (boost_cont_multialloc_arrays)
    ../vendor/simdjson/libsimdjson.a
                                  CMakeFiles/quick-lint-js.dir/main.cpp.o (_ZN8simdjson21active_implementationE)
    /usr/lib/x86_64-linux-gnu/libc_nonshared.a(elf-init.oS)
                                  /usr/lib/gcc/x86_64-linux-gnu/9/../../../x86_64-linux-gnu/Scrt1.o (__libc_csu_init)
    [snip]

The lines containing '../libboost.a' multiple times confuse our link map
parser. The confusion happens because there is no parenthesized symbol
name after the first mention of libboost.a. This means the following
spaces are considered part of the archive name, making the extracted
path "../libboost.a                 ../libboost.a ". This broken parsing
eventually leads to collect-copyright failing.

Fix our parsing by allowing the symbol name to be optional, and by
allowing a file-symbol pair to follow the main file.