Website: factor <title> and <meta description>
I want to use Open Graph annotations [1] on our web pages. In order to
do that, we need to duplicate the title and description in Open
Graph-friendly tags. Make this easy by putting the title and description
tags in common-head.ejs.html instead of in each page.

This commit should not change the generated HTML significantly.

[1] https://ogp.me/