Website: simplify BOM detection in error documentation
The error documentation highlights byte order marks (BOMs) by wrapping
them in a span. In order to detect a BOM, the code converts a string
into HTML then parses the HTML and checks its text content.

Parsing HTML is complicated and wasteful just to check for a BOM.
Simplify the code by checking for the BOM on the input string before
conversion into HTML.

This commit should not change behavior.