Create MSIX Windows installer
The winget package manager requires installers. It cannot install .exe
files itself. Create an MSIX-based installer for quick-lint-js. This
MSIX installer will be used in the future to add quick-lint-js to
winget.

Also, add code to sign a MSIX file. This requires a lot of code because
Go's archive/zip package produces ZIP files which can't be read by
Microsoft's APPX/MSIX installer stuff for some reason, and because we
need to recreate AppxBlockMap.xml when swapping out quick-lint-js.exe.