Refactor: flatten filesToTransform map-inside-map
filesToTransform is a nested map. This means that code signing can only
happen for archive mebmers. We can't code-sign a naked .exe, and we
can't code-sign an .exe inside a .zip inside a .zip. This limitation
prevents us from signing Chocolatey packages the way I want.

Make it easier (but not yet possible) to sign nested members by
flattening filesToTransform. Encode the nested paths in a DeepPath
struct.

This commit should not change behavior.