Refactor: allow signing files inside .zip-s inside .zip-s
If a .zip is inside a .zip, sign-release.go can't sign/transform files
in the inner .zip. This is because the .zip check only happens when
transforming regular files, not archive members.

Move the .zip-handling code to run for archive members and for regular
files. Do the same with the .tar.gz-handling code.

As a side effect, this change allows signing files outside any .zip or
.tar.gz. This is an added bonus, but I don't plan on using this
capability soon.