Refactor: use Relic instead of osslsigncode for Windows signing
We currently sign Windows .exe files using osslsigncode. I want to
create a MSIX installer for quick-lint-js for easier integration into
winget. osslsigncode does not support signing, MSIX installers, but
Relic does.

Switch from osslsigncode to Relic for signing and verifying Windows .exe
files. The private key used by Relic needs to be in a different format,
so document how to create a file in that format.

Also, add timestamp signature verification. (I think this was possible
with osslsigncode, but we didn't enable it. Relic fails all verification
if it can't verify the timestamp.)