Release: allow signing .zip and .tar.gz files directly
I want to implement signing for MSIX installers. MSIX installers are
ZIP-based, so it's possible we want to sign an executable inside the ZIP
while also signing the ZIP itself. Allow this kind of signing.

Because nothing currently is signed in this way, this commit should not
change behaviour. Future patches will use this new feature.