collect-copyright: refactor detection for special-cased .a/.lib
The archive_to_vendor_project dict serves two purposes:

1. map .a/.lib files to projects listed in vendor/README.txt
2. ignore certain .a/.lib files

The second purpose is implemented in a hacky way.

Refactor the code so that purpose 2 is no longer special; replace string
values in the map with objects with 'print_copyright' methods.
'print_copyright' can be customized per-.a/.lib.

This refactor will allow us to provide more detailed copyright
information, such as copyright for MSVC's statically-linked C++ library,
in the future.