Build: fix compiling node.def on Windows with Ninja
When building with Ninja, the CMAKE_VS_PLATFORM_NAME variable is not
set. This means /MACHINE isn't given to the 'lib' command, leading to
a warning that /MACHINE is not present. The warning leads to a build
failure because of /WX.

If CMAKE_VS_PLATFORM_NAME is missing, use CMAKE_SYSTEM_PROCESSOR
instead.