collect-copyright: refactor path to vendor directory
In collect-copyright, we currently store the path to 'vendor/' in a
global variable. This prevents us from support multiple vendor README
files, which is something I want to allow in the future.

Make the 'vendor/' path a property of each VendorReadmeProject, rather
than a global property.

This commit should not change behavior.