Fix debug code being included in Windows release builds
For some reason, GCC MinGW does not dead-code-strip most of
vector-profiler.cpp (and some other code) from quick-lint-js.exe. This
causes the .exe file to be bigger than it needs to be.

Move vector-profiler.cpp and debug-printing code into its own CMake
library which is not included by the quick-lint-js executable target.
This gives the linker no chance to accidentally include the vector
profiling and ostream code in release builds.

This commit should not change behavior.

This commit reduces the size of our production Windows binaries by a
little bit:

 file                                        |  before |   after | change
---------------------------------------------+---------+---------+--------
 manual/windows-arm.zip::quick-lint-js.exe   |  954368 |  954368 |     0%
 manual/windows-arm64.zip::quick-lint-js.exe | 1054208 | 1054208 |     0%
 manual/windows-x86.zip::quick-lint-js.exe   | 2092558 | 2020878 |    -3%
 manual/windows.zip::quick-lint-js.exe       | 1985536 | 1921024 |    -3%