Refactor: fix possible out-of-bounds read when reporting errors
vscode_error_formatter::append_diagnostic assumes that the error code
string_view is 0-terminated. It currently is 0-terminated, but this
relying on this is a bad idea.

Copy the error code into an std::string to force null termination. (We
were already creating a std::string anyway for other purposes.)

This commit should not change behavior.