Fix #44 to Warn if Integer Would Lose Precision
This pull request implements a new warning which warns if an integer literal is used which cannot be accurately represented by JavaScript's "Number" type and would lose precision. This is done by checking the number of digits in the number and using std::stold. Future implementations could expand this to warn about decimal numbers and numbers in binary/octal/hexadecimal.