fix(website): fix loading dusty.svg on Windows
On Windows, dusty.svg is a symbolic link pointing to another file. In my
Git checkout, the core.symlinks Git config is set to false, making Git
represent symbolic links as text files containing the target's path, not
as a copy or as a Windows link.

Work around Git's symlink behavior by following the symlink manually if
needed.