fix(website): fix &lt; in error list
The error list (https://quick-lint-js.com/errors/) shows the titles of
error files from Markdown. Some titles contain HTML or other Markdown,
causing titles such as the following to be shown to the user:

* E0187: mismatched JSX tags; expected &lt;/foo>
* E0195: missing parentheses around operand of `typeof`; `typeof`
  operator cannot be used before `**` without parentheses

Fix the parsing of HTML and Markdown by HTMLifying the titles before
storing into titleErrorDescription. Rename titleErrorDescription to
titleErrorDescriptionHTML to make it clearer that it now stores HTML.

This commit should not affect individual error pages which already
HTMLified the title.