refactor(arch): deduplicate GPG key in repo
dist/arch/quick-lint-js.gpg.key is [almost] a copy of
dist/certificates/quick-lint-js.gpg.key. Updating the file is
error-prone because both copies need to be updated.

Change dist/arch/quick-lint-js.gpg.key into a symlink, making updating
the GPG key easier.

Due to the symlink, we need to use 'tar' to build the Docker context.
It's gross, but it works.