fix(check-links): look at sources of <script>-s
check-links' logic for <script>-s is weird:

1. Find all <script> elements in an HTML page.
2. If all <script> elements have an src attribute, add those src
   attributes to the queue.

The "if" condition in step 2 above makes no sense to me. It causes most
pages to have their <script> elements unchecked, because most pages have
an inline script.

Fix the logic to enqueue the src of all <script> which have a src
attribute.