refactor(test): make interfaces_are_ignored_in_runtime_expressions more general
The interfaces_are_ignored_in_runtime_expressions test is hard-coded to
declaring the run-time variable with 'let'. I want to test declaring the
run-time variable with 'const' or 'function' too, but this means
diagnostics might change.

Refactor the diagnostics assertions to be possibly dependant on the
kind of the run-time variable.

This commit should not change test coverage.