feat(trace): create binary log format
I want to log document changes in the Visual Studio Code extension for
later analysis. Make this possible by designing a binary log format and
creating a reader and writer for this format.

This new code is not integrated with the VS Code extension or any other
part of quick-lint-js. Future commits will add integrations.