perf(website): avoid copying entire heap each keystroke
Uint8Array.slice makes a copy of the ArrayBuffer. We are slicing the
WebAssembly heap with only a start index, causing much of the heap to be
copied. Worse, we are doing this for each diagnostic! This slows down
linting significantly in the web demo.

Fix the bug by using subarray instead of slice. subarray does not copy
the underlying ArrayBuffer.

Credit to john_hillington_ for noticing the performance issue, and to
HPWebcamAble for finding the slice slowness.