refactor: simplify find_modifier; remove non-modifiers
When parsing class members, we store modifier keywords in a 'modifiers'
vector. Sometimes, it turns out a modifier isn't actually a keyword and
is instead the name of a method or field. In these cases, we compare the
stored modifier's span with the property name's span to see if the
modifier is a keyword or not.

Switch to a simpler design: When we determine that a modifier is the
property's name, remove the modifier from hte modifiers vector.