refactor: declare class statement name after }
In order to support TypeScript generics, we need to put class 'extends'
clauses inside the class's scope. However, moving the 'extends' clause
into the scope is impossible because visit_variable_declaration for the
class name needs to follow the 'extends' clause:

    class C/*2*/
        extends C/*1*/
        {/*3*/
            field;/*4*/
        }/*5*/

    1. visit_variable_use
    2. visit_variable_declaration (causes use-before-declaration error)
    3. visit_enter_class_scope
    4. visit_property_declaration
    5. visit_exit_class_scope

Maintain the ordering between the 'extends' clause and the class name by
moving the class name declaration after the class scope;

    class C/*5*/
        extends C/*1*/
        {/*2*/
            field;/*3*/
        }/*4*/

    1. visit_variable_use
    2. visit_enter_class_scope
    3. visit_property_declaration
    4. visit_exit_class_scope
    5. visit_variable_declaration (causes use-before-declaration error)

This change is only necessary for class statements, not for class
expressions. Class expressions include the class name inside the scope:

    (class C/*3*/
        extends C/*2*/
        {/*1*/
            field;/*4*/
        }/*5*/)

    1. visit_enter_class_scope
    2. visit_variable_use
    3. visit_variable_declaration (causes use-before-declaration error)
    4. visit_property_declaration
    5. visit_exit_class_scope

For consistency, we could declare class properties before the class name
in class expressions, but this patch doesn't do that.