fix(parse): fix use-after-free after rolling back transaction
Rolling back a transaction queues memory to be rewound. This is
incorrect if the transaction object is destructed later. This leads to a
use-after-free bug with the following input:

    await (0scribeBlock => {
        switch (child.type) {
        case 'test': {
        }
        }
        async describeBlock => {
        await (x) => {
        await (0scribeBlock => {
            await (x) => {
            await (0s => {
                @ @ @;
                @ @ @;

The specific cause is the following code pattern in
parser::parse_await_expression:

    {
      parser_transaction transaction = this->begin_transaction();
      // ...
      this->roll_back_transaction(std::move(transaction));

      this->diag_reporter_->report(diag_await_followed_by_arrow_function{
          .await_operator = operator_span,
      }); // (a)

    } // (b)

Line (a) allocates memory in the to-be-unwound allocator. Line (b)
performs the rewind. After (b), reading from diag_reporter_ (which in
the buggy case is a buffering_diag_reporter) leads to a use-after-free
error.

Fix the problem by having roll_back_transaction rewind immediately,
instead of deferring rewind to an arbitrary point in the future.

Fixing this issue exposed another use-after-free bug, causing
test_lex.rolled_back_outer_transaction_discards_errors to fail in ASAN
builds. Fix this by promptly destructing the buffering_diag_reporter
when committing a transaction.

Ref: 51a2ba73dae25dfb99d4460e3462f59605e4463e