refactor(parser): keep track of object literal init
The following two expressions have the same representation in
quick-lint-js but have different syntax rules. In particular, the former
is legal in a statement or in a structured binding, but the latter is
only allowed in a structured binding.

    {banana: orange = 42}
    {banana = 42}

Change the representation of object literals in our expression AST to
track the initializer/default separate from the value.