feat(typescript): parse type annotations in expressions
'x: T' in an expression looks like a type annotation but is invalid
TypeScript. Report a diagnostic for this code.

However, 'x: T' is allowed in arrow function headings. Therefore, parse
it as a new type of expression.

This commit does not implement 'x: T' in arrow function parameter lists.