refactor(typescript): use buffering_visitor instead of AST for types
Currently, we use an expression::variable for storing the type in a
TypeScript type-annotated expression (e.g. '(x: T)'). This is a hack,
and the original plan was to create a type expression AST hierarchy
similar to the value expression AST hierarchy.

It turns out an AST is unnecessary. We can instead store a list of
visits.

In expression::type_annotated, use buffering_visitor instead of
expression* for the type expression.

Currently, the buffering_visitor's memory is freed only when parsing
finishes. We should eventually free it periodically to reduce the high
water mark of memory usage.