fix: report depth-exceeded error for 'await await await ...'
When parsing 'await await await ...', the parser will begin a parser
transaction. This means that diagnostics reported through
parser::diag_reporter_ will be buffered. This buffering causes the
depth-exceeded diag to be buffered, but because the transaction is
aborted (not committed), the diag is never reported to the user.

Force the diag to be reported to the user by reporting as if no
transaction was active.

Also do the same for other diagnostics we report before longjmp-ing or
crashing (untested).