refactor(website): rewrite custom framework
The quick-lint-js website is built using a custom framework. It grew
organically and is pretty crusty. The dev server (run-server.mjs) and
static build (run-build.mjs) occasionally have different behavior, which
concerns me. Adding features requires changing both. And the code is
hard to follow.

Rewrite the framework on top of a VFS abstraction. Both the dev server
and static build use this abstraction to understand the directory
hierarchy. This should reduce divergence between the two modes.

The VFS in the new design is more complicated than the router in the old
design, but the dev server and static build code are much simpler.
Overall, there are fewer lines of production code with the new
framework.

Testing:

* The dev server tests show that the dev server behaves the same with
  the new framework as with the old framework.
* The static build produces the same output with the new framework as
  with the old framework.