feat(cli): report warning for vim-file-bufnr no file
outputs a warning if a --vim-file-bufnr flag was specified without a
file name or --stdin following it

fixes #201

I also did notice that the output_stream wouldn't be displayed if it
only contains warnings in lsp-mode. To fix this I added a `.flush()` at
the end of `options::dump_errors`.