fix(typescript): don't emit spurious type uses for x<y>z
An expression like 'x<y>z' should have three variable uses (x, y, and
z). In TypeScript mode, quick-lint-js reports an additional type
reference for 'y' because we parse it twice (once as a TypeScript
generic argument list, and once as a normal expression).

Don't emit a type use if parsing determines that '<y>' is not a generic
argument list.