refactor(test): create failing_diag_reporter for tests
When using the top-level parser_and_visit_statement function,
if parsing fails, gtest reports an error in parse-support.h with a
message including what diagnostics were reported. This is okay, but it's
sometimes hard to figure out what the state was when the diagnostic was
reported.

Instead of capturing errors and failing the test after parsing, fail the
test immediately. This means you can use --gtest_break_on_failure and
inspect parser state with a debugger.