fix: fix text desync issue for LSP and VS Code
If the user merges two lines, and the first line contains non-ASCII
characters, we incorrectly mark the merged line as containing only ASCII
characters. This causes later changes to the line to edit the wrong
parts of the line, leading to false positive errors and general
confusion.

The line-merging logic already considers the ASCII-ness of the bottom
line. Fix the logic to consider the ASCII-ness of the top line too.

This bug was discovered by T0x3139 and MiraBellier, and was fixed thanks
to a trace file provided by T0x3139.