@@@ refactor(lex): factor mask bit iteration into helper class
@@@ check x86 perf

parse_block_comment manually iterates over a mask bit-by-bit. This leads
to a lot of code inline, obscuring the algorithm. The algorithm is also
copy-pasted.

Simplify the code by factoring the bit scanning into a helper class.

In the future, bit scanning can be optimized [1], so making a separate
class makes optimizing easier.

[1] https://github.com/quick-lint/quick-lint-js/issues/798