perf(website): inline icon SVG files into HTML
We have some SVG files which are super tiny and are only referenced on
the home page. Embed them into the home page, possibly speeding up load
times.

The svg-sprite package is overkill for this task. However, we might use
svg-sprite in the future for other things, so I'm fine taking on the
dependency.