feat(website): reload dev server on .mjs changes
If a contributor changes certain .mjs files (e.g.
public/website/index.mjs), the web server ('yarn start') needs to be
restarted for the changes to take effect. This is due to Node.js' module
cache.

Restarting the server is inconvenient. We already have a mitigation
(website/src/render-ejs-file-child-process.mjs), but it doesn't work for
some cases.

Wrap our web server in a program which starts a new web server if a JS
(or WASM) file changes.