fix(website): fix macOS icon in dark mode
In dark mode, we filter the macOS icon to make it white. Our filter's
CSS selector doesn't match anymore because we put the macOS icon in a
sprite sheet.

Annotate images with their original icon using an HTML attribute (e.g.
data-icon="macos"), and use this attribute in the CSS selector.