fix(ci): reduce flakiness waiting for server to start
The CI script waits half a second for the development HTTP server to
start. CI runners might be slow, so half a second might not be enough
time. This can lead to spurious CI failures.

Instead of waiting half a second, try to connect to the HTTP server
until connecting succeeds. This should fix the flakiness, and probably
makes CI slightly faster too.