fix(typescript): don't double-warn for types in JavaScript
For the following JavaScript code, quick-lint-js reports multiple
diagnostics:

    type A = { k: number };

This is because type aliases are not supported in JS, and also type
annotations are not supported in JS. However, the type annotation is
inside a type, so it doesn't make sense to complain there.

Fix the diagnostics so only the type alias is diagnosed.