fix(fe): parse 'await < x' correctly in top level code
'await < x' in top level code should be treated as '<' applied to
'await' and 'x'. quick-lint-js first treats '<' as JSX and speculatively
parses. During this speculative parse, QLJS_PARSER_UNIMPLEMENTED is
called, causing all of parsing to stop.

Make the speculation fail when QLJS_PARSER_UNIMPLEMENTED is called, and
continue parsing '<' as a binary operator instead.