feat(typescript): parse optional parameters
Parse TypeScript optional parameters: '?' after parameter names. Create
an expression kind called 'optional' to store 'paramName?'.

Parsing is complicated by the fact that '(x ? : y)' is invalid (missing
expression between '?' and ':') but '(x ? : y) => z' is valid. To
disambiguate, we say that '? :' is parsed specially in parenthesized
expressions and in function parameter lists.