fix(fe): fix assertion failure with buffering_visitor_stack
If QLJS_PARSER_UNIMPLEMENTED is called, and it is caught, and there is a
stacked_buffering_visitor object somewhere in the call stack between the
call and the catch, then buffering_visitor_stack's stack gets in an
unexpected state. This causes an assertion failure.

Technically the assertion failure is harmless. However, disabling the
assertion sounds like a bad idea.

Make buffering_visitor_stack tolerant to caught
QLJS_PARSER_UNIMPLEMENTED-s by making it pop until the expected visitor
is found.