feat(typescript): don't treat eval specially in TypeScript code
In TypeScript, even though 'eval' can technically declare variables, the
TypeScript compiler complains about the following code:

    eval("var x = 42");
    x = 10;
    console.log(x);

Add a mode to the variable analyzer to not treat eval specially, making
quick-lint-js behave like the TypeScript compiler.

This mode is not enabled in the CLI or by any editor plugins yet.