feat(typescript): error when deleting any variable
In JavaScript, deleting variables is not an error. Deleting local
variables is likely a bug, but deleting global variables is possible.

In Typescript, deleting variables is always an error. TypeScript rejects
'delete someVar' syntactically, whether or not someVar exists or is a
global.

In TypeScript mode, report a new diagnostic if a variable is deleted.