feat(typescript): improve overload name mismatch diag (E0316)
E0316 reports an error if an adjacent pair of overloads have mismatched
names. This leads to confusing diagnostics if a function's middle
overload is misspelled:

    // error: function overload signature must be named 'key'
    function keys<T>(a: T[]): number[];
    // error: function overload signature must be named 'keys'
    function key(a: object): string[];
    function keys(a: object) {
        ...
    }

Fix this by checking overload names with the last overload rather than
checking adjacent overloads. The code is more complex for sure, but the
improved diagnostic is worth it.