feat(debug): create basic HTTP server
We have a vector profiler, but it only really works when using the CLI.
I want to use the vector profiler with quick-lint-js integrated into an
editor (e.g. when quick-lint-js running as an LSP server).

Make interactive profiling possible by adding an HTTP server to the LSP
server. The HTTP server does nothing interesting; it just serves a
'hello world' page. Future commits will add code to stream vector
profiler information.

The HTTP server is built using the Mongoose library.

This feature is opt-in at compile time for now.