refactor(logging): decouple file I/O from trace_flusher
trace_flusher is hard-coded to create files in a directory. I want to
add a different backend to trace_flusher.

Make future features in trace_flusher easier by decoupling trace_flusher
from the existing directory implementation. The interface of
trace_flusher is still the same, but the internals are factored.