fix(logging): fix wrong thread ID when enabling after registering
The tracing stream header contains a thread ID. When tracing is enabled,
then a thread is registered, the thread ID is correct. However, if a
thread is registered, then tracing is enabled, the thread ID is always
the thread which enabled (rather than the thread which registered).

Fix this by remembering the ID of the registered threads, and writing
that remembered ID rather than always the current thread ID.