refactor(logging): rewrite how thread header is sent to backend
Currently, when a new thread is registered or when a backend is enabled,
trace_flusher sends the thread header to the backend by writing data to
the thread's queue. This is bad if the queue has data already, which
will likely happen when we support multiple backends in the future.

Change the code to instead use a temporary queue, preventing any
interference between backends.