fix(test): fix flaky test
The test_trace_flusher.unregistering_thread_calls_thread_end test is
flaky. The test does not enforce that register_current_thread is called
on the main thread first, so threads 1 and 2 can be swapped. This
confuses our assertions which assume that thread 1 is the main thread.

Fix the test by registering the main thread before starting thread 2.