refactor(logging): rewrite trace-stream-reader to be incremental
read_trace_stream assumes that we have all the data available
immediately. This is not the case in some of our tests, making our test
code really hacky to work around this limitation.

Rewrite read_trace_stream to be incremental. Because state needs to be
maintained, make it a class called trace_stream_reader.

This commit does not clean up the aforementioned test code.