feat(cli): don't print escape sequences in dumb terminals
Just check for "dumb" in TERM. Trivial but works for most cases
where it's not appropriate to emit special codes by default, a
more roboust implementation would check for a capability in $TERM
(terminfo(5)). This simple boolean approach is also currently
used in gcc[1], which indicates that it's maybe good enough.

1: https://gcc.gnu.org/git?p=gcc.git;a=blob;f=gcc/diagnostic-color.cc;h=5f2856d18e48a642ef8646baa41a130926d8f387#l223

Signed-off-by: wagner riffel <w@104d.net>