refactor(lsp): allow lsp_linter to send any number of LSP messages
lsp_linter is forced to send exactly one LSP message. This restriction
is arbitrary, and we have an ugly workaround because of this.

Give lsp_linter::lint_and_get_diagnostics_notification an
outgoing_lsp_message_queue instead of a output byte_buffer, allowing an
lsp_linter to send zero messages if it wants.