fix(typescript): parse 'function f(p: (number)[]) {}' correctly
If a function parameter or return type starts with '(', then we parse
that type specially. The problem is that parsing '[]' or '|' after the
')' is skipped, leading to false positive errors.

Fix this by treating '('-starting types less specially, opting to call
parse_and_visit_typescript_type_expression (which handles '[]' and '|'
and such after the type) rather than just
parse_and_visit_typescript_arrow_or_paren_type_expression (which only
parses the parenthetical). (Parsing still needs to be special-cased for
return types in order to disambiguate some edge cases, but it doesn't
need to be so special that trailing '[]' parsing is skipped.)