refactor(tools): rewrite check-links in JavaScript/Node.js
People hacking on the website use Node.js. Currently they need to
install and configure another toolchain (Python and pip) in order to run
the check-links script. This is unnecessary friction.

Rewrite the check-links script in JavaScript+Node.js, reducing
dependencies for website contributors. The rewrite is mostly a
line-by-line port. Hopefully all features and bugs are faithfully
preserved.

Also, add some tests to increase confidence that the port works as
intended.