fix(ci): fix MinGW build error: "ninja: fatal: invalid -j parameter"
Rebuilding the quick-lint-js-github-mingw-builder Docker container fails with
the following message:

    ninja: fatal: invalid -j parameter
    The command '/bin/sh -c ./build-llvm.sh /opt/llvm-mingw' returned a non-zero code: 1

This is because the CORES variable is present but empty, causing the
build-llvm.sh script to pass '-j' (with no number) to Ninja. This is invalid.

Fix the build by removing the CORES variable entirely. I don't know why it was
there in the first place...