feat(build): generate lex-keyword-generated.cpp during builds
If lex-keyword-generated.cpp's generator
(tools/generate-lex-keyword.cpp) changes, or if the hash function (in
src/quick-lint-js/fe/keyword-lexer.h) changes, or if the list of
keywords (src/quick-lint-js/fe/keyword-list.h) changes,
lex-keyword-generated.cpp remains out of date unless you re-run the
quick-lint-js-generate-lex-keyword tool manually. This is hardly
documented, meaning "works on my machine" bugs can creep in.

Instead of making quick-lint-js-generate-lex-keyword a manually-executed
tool, run it as part of the build. For now, only do this when not
cross-compiling (because the quick-lint-js-generate-lex-keyword
executable is specific to the target machine and OS).

Also, fix tools/generate-lex-keyword.cpp to make it deterministic. (It
generated different output on my Linux machine than on my macOS
machine.)