feat(cli): list running quick-lint-js debug servers with --debug-apps
Add a new CLI option called --debug-apps. It lists all running instances
of quick-lint-js if the instance is running the debug server. For
example:

    $ ./build/quick-lint-js --debug-apps
    PROCESS ID  SERVER URL
    59558       http://localhost:8098

Currently there can be at most one debug app running (because the port
is hard-coded to 8098, and multiple apps cannot run on the same port).
This will be fixed in a future commit.

This feature is only implemented on Linux, macOS, and Windows.