feat(lex-tables): optionally use state machine tables to recognize some symbols
We use if-heavy code to lex symbols such as '>>>' and '!='. This code is ugly
and possibly slow.

Rewrite the parsing of these symbols to use a state machine table instead,
roughly similar to what code generated by the flex tool [1] does but using
different techniques.

With this patch, performance is about the same, unfortunately (<1% difference
parsing jQuery). Primarily, I suspect that parsing these symbols is not a
bottleneck. However, I think one thing holding this patch back performance-wise
is that the first character needs to be loaded and dispatched twice:
once for the switch (unchanged by this patch) and once for the initial
state in the table (added by this patch). In the future I hope to get
rid of this first load and dispatch, but I'm making no promise.

Note from the future: As I add more features to the lexer table,
performance gets worse. I haven't finished yet, so I am not able to test
the double-load double-dispatch theory. Therefore, I am leaving the
lexer tables as a disabled-by-default feature in case we want to easily
revert back.

The transition table and character class table are both hard to maintain by
hand. I hope to make a generator for these tables soon.

[1] https://github.com/westes/flex