refactor(lex-tables): generate lex tables instead of crafting by hand
The lex tables in src/quick-lint-js/fe/lex-tables.h are hand-crafted.
Adding new symbols is time-consuming, error prone, and not obvious.

Generate the lex tables from a list. This is overall more complex than a
hand-maintained table, but I think the complexity might be worth it in
the long run.