refactor(lex-tables): increase lex table state size
I want to add more handlers, but to do this, I need more bits in the
state. Add more bits by widening the state type from 8 bits to 16 bits.

In the future, I plan to reduce the number of data bits from 5 to 4, in
which case we can undo this 8->16 change.