refactor(lex-tables): make separate classification table for initial byte
The number of character classes is getting big, but many of the new ones
are not used in the transition table. To allow us to reduce the size and
maintenance cost of the transition table, make the initial character
class table possibly different. (With this patch, the two character
class tables have the same data, but this will change in the future.)