feat(diagnostic): add nullish_coalescing_operator diagnostic
This new diagnostic feature is a warning regarding
an expression. Typically, only variables or function
call return values have the ability to be null in
Javascript. When parse-expression recognizes a
binary operator, we report the diagnostic when it
is the nullish operator and the left child/children
are not a variable or function call."