feat(diagnostic): add nullish_coalescing_operator diagnostic
This new diagnostic feature is a warning regarding
an expression. The nullish coalescing operator
resolves to the right operand when the left operand
is null. However, if the left operand is never null,
it is pointless since the overall expression (LHS ??
RHS) will always resolve to the left operand.
Typically, only variables or function call return
values have the ability to be null in Javascript.
When parse-expression.cpp recognizes a binary
operator, we check that the operator is the nullish
operator and that the left child/children are not a
variable or function call and report the diagnostic.