fix(fe): require semicolon after async arrow function statements
quick-lint-js incorrectly reports that the 'else' is not attached to an
'if' in the following code:

    if (cond)
      async () => {};
    else {
    }

The problem was that the ';' was not interpreted as part of the
'async () => {}' statement. quick-lint-js thought the ';' was in between
the 'if' statement ('if (cond) async () => {}') and garbage ('else {}').

Consume semicolons after async arrow function statements, fixing parsing
of the above code.