fix(fe): require semicolon 'else (foo)'
quick-lint-js incorrectly reports that the 'do' is not attached to the
'while' in the following code:

    do
      if (c) {}
      else (b);
    while (false);

The problem was that the ';' was not interpreted as part of the
'(b);' statement. quick-lint-js thought the ';' was in between
the 'if-else' statement ('if (c) {} else (b)') and a 'while' loop
('while (false);'').

Consume semicolons after 'else (...)', fixing parsing of the above code.