fix(diag) add mistyped_strict_inequality_operator
This diagnostic is used to warn if a null assertion operator in TypeScript, `! ==`, was intended to be a strict inequality operator `!==`, instead. Will trigger anytime an equals_operator token follows a bang token.

Note: Not able to pass tests when they are incased in an if-statement, but with said if statements removed, all tests pass fine.
Also still needs formatting fixes.