Fix(diag) Now prioritizes E0373 over E0261
The code will now raise the warning regardless of if the code is JavaScript or TypeScript, and prioritizes raising this warning over E0261: typescript_non_null_assertion_not_allowed_in_javascript.
Additionally, I made the code examples in the documentation into TypeScript, since I forgot to in the first place.

Let me know if this is now ready to land.